+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v10.4.0
 
 ### New Features
-- Added helper for parsing Azure Resource id's.
+- Added helper for parsing Azure Resource ID's.
 - Added deprecation message to utils.GetEnvVarOrExit()
 
 ## v10.3.0